backend/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Cors==5.0.1
Werkzeug==3.1.3
Flask-JWT-Extended==4.7.1
python-dotenv==1.1.0